+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## 2.10.0
+
+* Add repo LWRP for interfacing with stash repositories
+
+## 2.9.0
+
+* Add hook LWRP for interfacing with stash hooks
+
 ## 2.8.0
 
 * Checksum attribute auto-detection